Timer is double now

+++ CodingDojo/games/kata/src/main/java/com/codenjoy/dojo/kata/model/Timer.java
@@ -41,8 +41,8 @@ public class Timer {
         return Calendar.getInstance().getTimeInMillis();
     }
 
-    public int end() {
+    public double end() {
         long delta = now() - start;
-        return (int)(delta / ONE_MINUTE_IN_MILLS);
+        return (double)(delta / ONE_MINUTE_IN_MILLS);
     }
 }
